refactor: chat[name] 페이지 라우팅 chat으로 수정
