+++ apps/frontend/src/_components/pages/chat/RecordButton/RecordButton.tsx
