+++ apps/frontend/src/app/api/chat/route.ts
@@ -24,4 +24,4 @@ export async function POST(req: Request) {
       return new Response(error.message, { status: 500 });
     }
   }
-}
+}
