+++ apps/frontend/src/app/chat/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import InterviewerProfileSection from "@/_components/pages/chat/[name]/InterviewerProfileSection";
+import InterviewerProfileSection from "@/_components/pages/chat/InterviewerProfileSection";
 import styles from "./page.module.css";
 import Text from "@repo/ui/Text";
-import ChatSection from "@/_components/pages/chat/[name]/ChatSection";
+import ChatSection from "@/_components/pages/chat/ChatSection";
 import useUserStore from "@/_store/useUserStore";
 import { useInterviewerStore } from "@/_store/useInterviewerStore";
 import { useRouter } from "next/navigation";
